+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/schema/pg/AbstractPgTable.java
@@ -50,7 +50,6 @@ public abstract class AbstractPgTable extends AbstractTable {
 
     protected boolean hasOids;
     protected final List<Inherits> inherits = new ArrayList<>();
-    private String method = Consts.HEAP;
 
     protected AbstractPgTable(String name) {
         super(name);
@@ -63,7 +62,6 @@ public abstract class AbstractPgTable extends AbstractTable {
         appendName(sbSQL);
         appendColumns(sbSQL, sbOption);
         appendInherit(sbSQL);
-        appendAccessMethod(sbSQL);
         appendOptions(sbSQL);
         sbSQL.append(sbOption);
         appendAlterOptions(sbSQL);
@@ -124,19 +122,6 @@ public abstract class AbstractPgTable extends AbstractTable {
             }
             sbSQL.setLength(sbSQL.length() - 2);
             sbSQL.append(")");
-        }
-    }
-
-    /**
-     * Appends table access method. <br>
-     * (defined here because the table access method is not used for partitioned and foreign tables)
-     *
-     * @param sbSQL
-     *            - StringBuilder for access method
-     */
-    protected void appendAccessMethod(StringBuilder sbSQL) {
-        if (!Consts.HEAP.equals(method)) {
-            sbSQL.append("\nUSING ").append(PgDiffUtils.getQuotedName(method));
         }
     }
 
@@ -208,10 +193,6 @@ public abstract class AbstractPgTable extends AbstractTable {
 
     @Override
     protected boolean isNeedRecreate(AbstractTable newTable) {
-        if (!Objects.equals(getMethod(), ((AbstractPgTable) newTable).getMethod())) {
-            return true;
-        }
-
         if (options.equals(newTable.getOptions())) {
             return false;
         }
@@ -325,15 +306,6 @@ public abstract class AbstractPgTable extends AbstractTable {
 
     public void setHasOids(final boolean hasOids) {
         this.hasOids = hasOids;
-        resetHash();
-    }
-
-    public String getMethod() {
-        return method;
-    }
-
-    public void setMethod(String using) {
-        this.method = using;
         resetHash();
     }
 
@@ -463,8 +435,7 @@ public abstract class AbstractPgTable extends AbstractTable {
             AbstractPgTable table = (AbstractPgTable) obj;
 
             return hasOids == table.getHasOids()
-                    && inherits.equals(table.inherits)
-                    && Objects.equals(method, table.getMethod());
+                    && inherits.equals(table.inherits);
         }
         return false;
     }
@@ -474,7 +445,6 @@ public abstract class AbstractPgTable extends AbstractTable {
         super.computeHash(hasher);
         hasher.putOrdered(inherits);
         hasher.put(hasOids);
-        hasher.put(method);
     }
 
     @Override
@@ -482,7 +452,6 @@ public abstract class AbstractPgTable extends AbstractTable {
         AbstractPgTable copy = (AbstractPgTable) super.shallowCopy();
         copy.inherits.addAll(inherits);
         copy.setHasOids(getHasOids());
-        copy.setMethod(getMethod());
         return copy;
     }
 }